Move Generator to 'dictionary'
